Feat(spark): new Spark2 dialect, improve DATEDIFF sql generation BREAKING
